+++ python/tests/operations/dot.py
@@ -2,11 +2,9 @@ import os
 import unittest
 
 import numpy as np
-from utils import compare_blocks
 
 import equistore.io
 import equistore.operations as fn
-from equistore.operations.dot import _dot_block
 from equistore import Labels, TensorBlock, TensorMap
 
 DATA_ROOT = os.path.join(os.path.dirname(__file__), "..", "data")
@@ -16,26 +14,29 @@ class TestDot(unittest.TestCase):
     def test_self_dot_no_components(self):
         tensor1 = equistore.io.load(
             os.path.join(DATA_ROOT, "qm7-power-spectrum.npz"),
+            # the npz is using DEFLATE compression, equistore only supports STORED
             use_numpy=True,
         )
         tensor2 = fn.remove_gradients(tensor1)
-        dot_blocks = []
+
+        dot_tensor = fn.dot(A=tensor1, B=tensor2)
+        self.assertTrue(np.all(tensor1.keys == dot_tensor.keys))
+
         for key, block1 in tensor1:
+            dot_block = dot_tensor.block(key)
+
             block2 = tensor2.block(key)
-            result_block = _dot_block(block1, block2)
-            dot_blocks.append(result_block)
             expected_values = np.dot(block1.values, block2.values.T)
-            self.assertTrue(
-                np.allclose(result_block.values, expected_values, rtol=1e-13)
-            )
-            self.assertTrue(np.all(block1.samples == result_block.samples))
-            self.assertTrue(np.all(block2.samples == result_block.properties))
+
+            self.assertTrue(np.allclose(dot_block.values, expected_values, rtol=1e-13))
+            self.assertTrue(np.all(block1.samples == dot_block.samples))
+            self.assertTrue(np.all(block2.samples == dot_block.properties))
 
             self.assertTrue(
-                len(block1.gradients_list()) == len(result_block.gradients_list())
+                len(block1.gradients_list()) == len(dot_block.gradients_list())
             )
             for parameter, gradient1 in block1.gradients():
-                result_gradient = result_block.gradient(parameter)
+                result_gradient = dot_block.gradient(parameter)
                 self.assertTrue(np.all(gradient1.samples == result_gradient.samples))
                 self.assertTrue(
                     len(gradient1.components) == len(result_gradient.components)
@@ -51,54 +52,53 @@ class TestDot(unittest.TestCase):
                 self.assertTrue(
                     np.allclose(expected_data, result_gradient.data, rtol=1e-13)
                 )
-        expected_tensor = TensorMap(tensor1.keys, dot_blocks)
-
-        dot_tensor = fn.dot(A=tensor1, B=tensor2)
-        self.assertTrue(np.all(expected_tensor.keys == dot_tensor.keys))
-        for key, expected_block in expected_tensor:
-
-            comparing_dict = compare_blocks(expected_block, dot_tensor.block(key))
-            self.assertTrue(comparing_dict["general"])
 
     def test_self_dot_components(self):
         tensor1 = equistore.io.load(
             os.path.join(DATA_ROOT, "qm7-spherical-expansion.npz"),
             use_numpy=True,
         )
+
         tensor2 = []
-        dot_blocks = []
-        isample = 1
+        n_samples = 42
         for key, block1 in tensor1:
-            value2 = np.arange(isample * len(block1.properties)).reshape(
-                isample, len(block1.properties)
+            value2 = np.arange(n_samples * len(block1.properties)).reshape(
+                n_samples, len(block1.properties)
             )
 
             block2 = TensorBlock(
                 values=value2,
                 samples=Labels(
-                    ["samples"], np.array([[i] for i in range(isample)], dtype=np.int32)
+                    ["samples"],
+                    np.array([[i] for i in range(n_samples)], dtype=np.int32),
                 ),
                 components=[],
                 properties=block1.properties,
             )
             tensor2.append(block2)
-            result_block = _dot_block(block1, block2)
-            dot_blocks.append(result_block)
+
+        tensor2 = TensorMap(tensor1.keys, tensor2)
+
+        dot_tensor = fn.dot(A=tensor1, B=tensor2)
+        self.assertTrue(np.all(tensor1.keys == dot_tensor.keys))
+
+        for key, block1 in tensor1:
+            block2 = tensor2.block(key)
+            dot_block = dot_tensor.block(key)
+
             expected_values = np.dot(
                 block1.values,
                 block2.values.T,
             )
-            self.assertTrue(
-                np.allclose(result_block.values, expected_values, rtol=1e-13)
-            )
-            self.assertTrue(np.all(block1.samples == result_block.samples))
-            self.assertTrue(np.all(block2.samples == result_block.properties))
+            self.assertTrue(np.allclose(dot_block.values, expected_values, rtol=1e-13))
+            self.assertTrue(np.all(block1.samples == dot_block.samples))
+            self.assertTrue(np.all(block2.samples == dot_block.properties))
 
             self.assertTrue(
-                len(block1.gradients_list()) == len(result_block.gradients_list())
+                len(block1.gradients_list()) == len(dot_block.gradients_list())
             )
             for parameter, gradient1 in block1.gradients():
-                result_gradient = result_block.gradient(parameter)
+                result_gradient = dot_block.gradient(parameter)
                 self.assertTrue(np.all(gradient1.samples == result_gradient.samples))
                 for i in range(len(gradient1.components)):
                     self.assertTrue(
@@ -115,15 +115,6 @@ class TestDot(unittest.TestCase):
                 self.assertTrue(
                     np.allclose(expected_data, result_gradient.data, rtol=1e-13)
                 )
-        expected_tensor = TensorMap(tensor1.keys, dot_blocks)
-        tensor2 = TensorMap(tensor1.keys, tensor2)
-
-        dot_tensor = fn.dot(A=tensor1, B=tensor2)
-        self.assertTrue(np.all(expected_tensor.keys == dot_tensor.keys))
-        for key, expected_block in expected_tensor:
-
-            comparing_dict = compare_blocks(expected_block, dot_tensor.block(key))
-            self.assertTrue(comparing_dict["general"])
 
 
 # TODO: add tests with torch & torch scripting/tracing
